🤖 Refactor(client): group 3d 변환 로직 아이소메트릭 변환 공식을 사용하여 변경
