+++ apps/client/src/components/Group/ncloud/Rect3D.tsx
@@ -1,45 +1,34 @@
-import { Bounds } from '@types';
-import { screenToGrid2d, gridToScreen3d } from '@utils';
-import { ReactNode } from 'react';
+import Text from '@components/Group/ncloud/Title';
+import { useDimensionContext } from '@contexts/DimensionContext';
+import { Bounds, Group } from '@types';
+import { generateRandomRGB } from '@utils';
+import { useMemo } from 'react';
 
-type Props = {
-    children: ReactNode;
-    bounds: Bounds;
+interface Props extends Partial<Group> {
     color: string;
-    [key: string]: any;
+    bounds: Bounds;
+}
+
+const Rect3D = ({ bounds, properties, color }: Props) => {
+    const transform = 'matrix(0.707 0.409 -0.707 0.409 0 0)';
+    const points = `90 0, 90 ${bounds.height}, ${bounds.width + 90} ${bounds.height}, ${bounds.width + 90} 0`;
+
+    return (
+        <>
+            <polygon
+                points={points}
+                stroke={color}
+                strokeWidth="8"
+                fill="none"
+                transform={transform}
+            ></polygon>
+            <Text bounds={bounds} color={color} text={properties?.name} />
+        </>
+    );
 };
-export default ({ bounds, properties, color, children }: Props) => {
-    const topLeftGrid = screenToGrid2d({ x: 0, y: 0 });
-    const topRightGrid = screenToGrid2d({ x: bounds.width, y: 0 });
-    const bottomRightGrid = screenToGrid2d({
-        x: bounds.width,
-        y: bounds.height,
-    });
-    const bottomLeftGrid = screenToGrid2d({ x: 0, y: bounds.height });
 
-    const point1 = gridToScreen3d({
-        col: topLeftGrid.col + 1,
-        row: topLeftGrid.row,
-    });
-    const point2 = gridToScreen3d({
-        col: topRightGrid.col + 1,
-        row: topRightGrid.row,
-    });
-    const point3 = gridToScreen3d({
-        col: bottomRightGrid.col + 1,
-        row: bottomRightGrid.row,
-    });
-    const point4 = gridToScreen3d({
-        col: bottomLeftGrid.col + 1,
-        row: bottomLeftGrid.row,
-    });
-
-    const points = `
-        ${point1.x} ${point1.y},
-        ${point2.x} ${point2.y},
-        ${point3.x} ${point3.y},
-        ${point4.x} ${point4.y}
-    `;
+const Rect2D = ({ bounds, color, properties }: Props) => {
+    const points = `0 0, 0 ${bounds.height}, ${bounds.width} ${bounds.height}, ${bounds.width} 0`;
 
     return (
         <>
@@ -49,7 +38,18 @@ export default ({ bounds, properties, color, children }: Props) => {
                 strokeWidth="8"
                 fill="none"
             ></polygon>
-            {children}
+            <Text bounds={bounds} color={color} text={properties?.name} />
         </>
+    );
+};
+
+export default ({ bounds, properties }: Omit<Props, 'color'>) => {
+    const { dimension } = useDimensionContext();
+    const color = useMemo(() => generateRandomRGB(), []);
+
+    return dimension === '2d' ? (
+        <Rect2D bounds={bounds} properties={properties} color={color} />
+    ) : (
+        <Rect3D bounds={bounds} properties={properties} color={color} />
     );
 };
